refactor: CohortRepository 분리
