+++ src/main/java/nextstep/courses/infrastructure/JdbcCourseRepository.java
@@ -46,29 +46,4 @@ public class JdbcCourseRepository implements CourseRepository {
                     rs.getTimestamp("updated_at").toLocalDateTime()
             ), id);
   }
-
-  @Override
-  public long saveCohort(CohortEntity cohortEntity) {
-    String sql = "INSERT INTO cohort (course_id, name) values(?, ?)";
-    KeyHolder keyHolder = new GeneratedKeyHolder();
-
-    jdbcTemplate.update(connection -> {
-      var ps = connection.prepareStatement(sql, new String[]{"id"});
-      ps.setLong(1, cohortEntity.courseId());
-      ps.setString(2, cohortEntity.name());
-      return ps;
-    }, keyHolder);
-    return keyHolder.getKey().longValue();
-  }
-
-  @Override
-  public CohortEntity findCohortByCourseId(Long courseId) {
-    String sql = "SELECT id, course_id, name FROM cohort WHERE course_id = ?";
-    return jdbcTemplate.queryForObject(sql, (rs, rowNum) ->
-            new CohortEntity(
-                    rs.getLong("id"),
-                    rs.getLong("course_id"),
-                    rs.getString("name")
-            ), courseId);
-  }
 }
