+++ src/main/java/nextstep/courses/repository/CohortRepository.java
@@ -1,0 +1,8 @@
+package nextstep.courses.repository;
+
+import nextstep.courses.entity.CohortEntity;
+
+public interface CohortRepository {
+  long saveCohort(CohortEntity cohort);
+  CohortEntity findByCourseId(Long courseId);
+}
