+++ src/main/java/nextstep/courses/repository/CourseRepository.java
@@ -1,11 +1,8 @@
 package nextstep.courses.repository;
 
-import nextstep.courses.entity.CohortEntity;
 import nextstep.courses.entity.CourseEntity;
 
 public interface CourseRepository {
   long save(CourseEntity course);
   CourseEntity findById(Long id);
-  long saveCohort(CohortEntity cohort);
-  CohortEntity findCohortByCourseId(Long courseId);
 }
