+++ src/main/java/nextstep/courses/service/CourseService.java
@@ -2,6 +2,7 @@ package nextstep.courses.service;
 
 import nextstep.courses.domain.Course;
 import nextstep.courses.entity.CohortEntity;
+import nextstep.courses.repository.CohortRepository;
 import nextstep.courses.repository.CourseRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,20 +12,22 @@ import org.springframework.transaction.annotation.Transactional;
 public class CourseService {
 
   private final CourseRepository courseRepository;
+  private final CohortRepository cohortRepository;
 
-  public CourseService(CourseRepository courseRepository) {
+  public CourseService(CourseRepository courseRepository, CohortRepository cohortRepository) {
     this.courseRepository = courseRepository;
+    this.cohortRepository = cohortRepository;
   }
 
   public long save(Course course) {
     long courseId = courseRepository.save(course.toCourseEntity());
     CohortEntity cohortEntity = course.cohort().toCohortEntity(courseId);
-    courseRepository.saveCohort(cohortEntity);
+    cohortRepository.saveCohort(cohortEntity);
     return courseId;
   }
 
   public Course findById(Long id) {
-    CohortEntity cohortEntity = courseRepository.findCohortByCourseId(id);
+    CohortEntity cohortEntity = cohortRepository.findByCourseId(id);
     return courseRepository.findById(id).toCourse(cohortEntity.toCohort());
   }
 }
